refactor: Moved NetworkedVar classes to own files
